requirements.txt
certifi==2022.6.15
charset-normalizer==2.1.0
idna==3.3
pyweatherbit==2.1.1
requests==2.28.1
responses==0.21.0
urllib3==1.26.11
